doc: add doxygen and format lexer_list.c
